state view model for redux and project. Duck/slice model for each property of the store
