+++ src/state/_shared/store/index.js
@@ -1,0 +1,7 @@
+// This is where we will store all the states
+
+export const apiInitialState = {
+    loading: false,
+    error: "",
+    searchValue: "",
+  };
