+++ src/state/actionTypes.js
@@ -1,1 +1,5 @@
 // file to store the action type constants
+export const FETCHING_API_START = "FETCHING_API_LOADING";
+export const FETCHING_API_SUCCESS = "FETCHING_API_SUCCESS";
+export const FETCHING_API_FAILURE = "FETCHING_API_FAIL";
+export const SEARCH_VALUE = "SEARCH_VALUE";
