--- src/state/index.js
@@ -1,16 +1,0 @@
-//this handles all initial state values
-
-
-// also imports all action creators, actions, and rootReducer.
-// other parts of the code can destructure from this file.
-
-// each reducer and their action creators should be included in a folder inside state 
-
-
-// -state
-// ----AState
-// --------actions.js
-// --------reducer.js
-// ----BState
-// --------actions.js
-// --------reducer.js
