+++ src/state/rootReducer.js
@@ -1,1 +1,3 @@
 // where we will combind all the reducers into a rootReducer
+import { combineReducers } from "redux";
+
