+++ src/view/components/About.js
