+++ src/view/components/Cocktail.js
