+++ src/view/components/index.js
